Mutable filter improved.

+++ src/test/scala/com/github/oskin1/scakoo/CuckooFilterTestHelper.scala
@@ -9,20 +9,20 @@ trait CuckooFilterTestHelper {
   val immutableFilterGen: Gen[immutable.CuckooFilter[Array[Byte]]] = for {
     entries <- Gen.chooseNum(1, 16)
     powerOfTwo <- Gen.chooseNum(8, 16)
-  } yield newImmutableFilter(entries, math.pow(2, powerOfTwo).toLong)
+  } yield newImmutableFilter(entries, math.pow(2, powerOfTwo).toInt)
 
   val mutableFilterGen: Gen[mutable.CuckooFilter[Array[Byte]]] = for {
     entries <- Gen.chooseNum(1, 16)
     powerOfTwo <- Gen.chooseNum(8, 16)
-  } yield newMutableFilter(entries, math.pow(2, powerOfTwo).toLong)
+  } yield newMutableFilter(entries, math.pow(2, powerOfTwo).toInt)
 
-  def newImmutableFilter(entriesPerBucket: Int, bucketsQty: Long): immutable.CuckooFilter[Array[Byte]] = {
+  def newImmutableFilter(entriesPerBucket: Int, bucketsQty: Int): immutable.CuckooFilter[Array[Byte]] = {
     import Funnel.byteArrayFunnel
     import TaggingStrategy.MurmurHash3Strategy
     immutable.CuckooFilter(entriesPerBucket, bucketsQty)
   }
 
-  def newMutableFilter(entriesPerBucket: Int, bucketsQty: Long): mutable.CuckooFilter[Array[Byte]] = {
+  def newMutableFilter(entriesPerBucket: Int, bucketsQty: Int): mutable.CuckooFilter[Array[Byte]] = {
     import Funnel.byteArrayFunnel
     import TaggingStrategy.MurmurHash3Strategy
     mutable.CuckooFilter(entriesPerBucket, bucketsQty)
